New Logger, Persistent location & World Bosses
Changed the logger to append to a file so it does not always control it.
The logger is now controlled through the main ui via a checkbox.
Starting the logger clears the old file. The location of the DPS
Display, Health Bar, and the Timer are are now stored and reloaded when
the gw2dps is restarted. Also added support for world bosses (Fixes #1).